+++ src/SdJwt.Net.Eudiw/Credentials/PidCredentialHandler.cs
@@ -3,145 +3,6 @@ using SdJwt.Net.Eudiw.Arf;
 namespace SdJwt.Net.Eudiw.Credentials;
 
 /// <summary>
-/// Exception thrown when PID credential validation fails.
-/// </summary>
-public class PidValidationException : Exception
-{
-    /// <summary>
-    /// Initializes a new instance of the <see cref="PidValidationException"/> class.
-    /// </summary>
-    /// <param name="message">The error message.</param>
-    public PidValidationException(string message) : base(message) { }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="PidValidationException"/> class.
-    /// </summary>
-    /// <param name="message">The error message.</param>
-    /// <param name="innerException">The inner exception.</param>
-    public PidValidationException(string message, Exception innerException) : base(message, innerException) { }
-}
-
-/// <summary>
-/// Result of PID credential validation.
-/// </summary>
-public class PidValidationResult
-{
-    /// <summary>
-    /// Indicates whether validation succeeded.
-    /// </summary>
-    public bool IsValid
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// List of validation errors.
-    /// </summary>
-    public IReadOnlyList<string> Errors { get; set; } = Array.Empty<string>();
-
-    /// <summary>
-    /// Creates a successful validation result.
-    /// </summary>
-    /// <returns>A successful PID validation result.</returns>
-    public static PidValidationResult Success()
-    {
-        return new PidValidationResult { IsValid = true };
-    }
-
-    /// <summary>
-    /// Creates a failed validation result.
-    /// </summary>
-    /// <param name="errors">The validation errors.</param>
-    /// <returns>A failed PID validation result.</returns>
-    public static PidValidationResult Failure(params string[] errors)
-    {
-        return new PidValidationResult { IsValid = false, Errors = errors };
-    }
-}
-
-/// <summary>
-/// Model representing PID credential data.
-/// </summary>
-public class PidCredential
-{
-    /// <summary>
-    /// Family name of the holder.
-    /// </summary>
-    public string FamilyName { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Given name of the holder.
-    /// </summary>
-    public string GivenName { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Birth date of the holder.
-    /// </summary>
-    public DateTime BirthDate
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Issuance date of the credential.
-    /// </summary>
-    public DateTime IssuanceDate
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Expiry date of the credential.
-    /// </summary>
-    public DateTime ExpiryDate
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Issuing authority name.
-    /// </summary>
-    public string IssuingAuthority { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Issuing country (ISO 3166-1 alpha-2).
-    /// </summary>
-    public string IssuingCountry { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Indicates if holder is over 18, if disclosed.
-    /// </summary>
-    public bool? AgeOver18
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Indicates if holder is over 21, if disclosed.
-    /// </summary>
-    public bool? AgeOver21
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Nationality of the holder, if disclosed.
-    /// </summary>
-    public string? Nationality
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Resident address, if disclosed.
-    /// </summary>
-    public string? ResidentAddress
-    {
-        get; set;
-    }
-}
-
-/// <summary>
 /// Handles PID (Person Identification Data) credential processing per EUDIW ARF.
 /// </summary>
 public class PidCredentialHandler
