+++ src/SdJwt.Net.Eudiw/EudiWalletOptions.cs
@@ -1,5 +1,3 @@
-using SdJwt.Net.Eudiw.Arf;
-using SdJwt.Net.Eudiw.TrustFramework;
 using SdJwt.Net.Wallet.Attestation;
 using SdJwt.Net.Wallet.Audit;
 using SdJwt.Net.Wallet.Protocols;
@@ -145,184 +143,4 @@ public class EudiWalletOptions
     {
         get; set;
     }
-}
-
-/// <summary>
-/// Result of EU Trust validation.
-/// </summary>
-public class EuTrustValidationResult
-{
-    /// <summary>
-    /// Whether the issuer is trusted.
-    /// </summary>
-    public bool IsTrusted
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Member state of the trusted service provider.
-    /// </summary>
-    public string? MemberState
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Service type of the provider.
-    /// </summary>
-    public TrustServiceType ServiceType
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Validation errors if not trusted.
-    /// </summary>
-    public IReadOnlyList<string> Errors { get; set; } = Array.Empty<string>();
-
-    /// <summary>
-    /// Creates a trusted result.
-    /// </summary>
-    /// <param name="memberState">The member state.</param>
-    /// <param name="serviceType">The service type.</param>
-    /// <returns>A trusted result.</returns>
-    public static EuTrustValidationResult Trusted(string memberState, TrustServiceType serviceType = TrustServiceType.QualifiedAttestation)
-    {
-        return new EuTrustValidationResult
-        {
-            IsTrusted = true,
-            MemberState = memberState,
-            ServiceType = serviceType
-        };
-    }
-
-    /// <summary>
-    /// Creates an untrusted result.
-    /// </summary>
-    /// <param name="errors">The validation errors.</param>
-    /// <returns>An untrusted result.</returns>
-    public static EuTrustValidationResult Untrusted(params string[] errors)
-    {
-        return new EuTrustValidationResult
-        {
-            IsTrusted = false,
-            Errors = errors
-        };
-    }
-}
-
-/// <summary>
-/// Exception thrown for EUDI Trust validation failures.
-/// </summary>
-public class EudiTrustException : Exception
-{
-    /// <summary>
-    /// Initializes a new instance of the <see cref="EudiTrustException"/> class.
-    /// </summary>
-    /// <param name="message">The error message.</param>
-    public EudiTrustException(string message) : base(message) { }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="EudiTrustException"/> class.
-    /// </summary>
-    /// <param name="message">The error message.</param>
-    /// <param name="innerException">The inner exception.</param>
-    public EudiTrustException(string message, Exception innerException)
-        : base(message, innerException) { }
-}
-
-/// <summary>
-/// Exception thrown for ARF compliance violations.
-/// </summary>
-public class ArfComplianceException : Exception
-{
-    /// <summary>
-    /// Gets the list of ARF violations.
-    /// </summary>
-    public IReadOnlyList<string> Violations
-    {
-        get;
-    }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="ArfComplianceException"/> class.
-    /// </summary>
-    /// <param name="violations">The list of ARF compliance violations.</param>
-    public ArfComplianceException(IReadOnlyList<string> violations)
-        : base($"ARF compliance violations: {string.Join(", ", violations)}")
-    {
-        Violations = violations;
-    }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="ArfComplianceException"/> class.
-    /// </summary>
-    /// <param name="message">The error message.</param>
-    public ArfComplianceException(string message)
-        : base(message)
-    {
-        Violations = new[] { message };
-    }
-}
-
-/// <summary>
-/// Options for PID credential request.
-/// </summary>
-public class PidRequestOptions
-{
-    /// <summary>
-    /// Authentication method to use for issuance.
-    /// </summary>
-    public EudiAuthMethod AuthMethod { get; set; } = EudiAuthMethod.AuthorizationCode;
-
-    /// <summary>
-    /// Pre-authorized code if using pre-auth flow.
-    /// </summary>
-    public string? PreAuthorizedCode
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// User PIN if required.
-    /// </summary>
-    public string? UserPin
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Key algorithm for credential binding.
-    /// </summary>
-    public string KeyAlgorithm { get; set; } = "ES256";
-
-    /// <summary>
-    /// Redirect URI for authorization code flow.
-    /// </summary>
-    public string? RedirectUri
-    {
-        get; set;
-    }
-}
-
-/// <summary>
-/// Authentication methods for EUDI credential issuance.
-/// </summary>
-public enum EudiAuthMethod
-{
-    /// <summary>
-    /// OAuth 2.0 Authorization Code flow.
-    /// </summary>
-    AuthorizationCode,
-
-    /// <summary>
-    /// Pre-Authorized Code flow.
-    /// </summary>
-    PreAuthorized,
-
-    /// <summary>
-    /// eIDAS authentication.
-    /// </summary>
-    Eidas
 }
