+++ src/SdJwt.Net.Wallet/Core/IKeyManager.cs
@@ -3,118 +3,6 @@ using Microsoft.IdentityModel.Tokens;
 namespace SdJwt.Net.Wallet.Core;
 
 /// <summary>
-/// Information about a managed key.
-/// </summary>
-public class KeyInfo
-{
-    /// <summary>
-    /// Unique key identifier.
-    /// </summary>
-    public string KeyId { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Algorithm the key is used for.
-    /// </summary>
-    public string Algorithm { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Key type (EC, RSA).
-    /// </summary>
-    public string KeyType { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Curve name for EC keys.
-    /// </summary>
-    public string? Curve
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// When the key was created.
-    /// </summary>
-    public DateTimeOffset CreatedAt
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Whether the key is hardware-backed.
-    /// </summary>
-    public bool IsHardwareBacked
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Secure area name if hardware-backed.
-    /// </summary>
-    public string? SecureAreaName
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// HAIP compliance level if validated.
-    /// </summary>
-    public int? HaipLevel
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Additional metadata.
-    /// </summary>
-    public IDictionary<string, object>? Metadata
-    {
-        get; set;
-    }
-}
-
-/// <summary>
-/// Options for key generation supporting HAIP compliance levels.
-/// </summary>
-public class KeyGenerationOptions
-{
-    /// <summary>
-    /// Signing algorithm (e.g., ES256, ES384, ES512).
-    /// </summary>
-    public string Algorithm { get; set; } = "ES256";
-
-    /// <summary>
-    /// Required HAIP compliance level (1, 2, or 3).
-    /// </summary>
-    public int? RequiredHaipLevel
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Whether to require hardware-backed key storage.
-    /// </summary>
-    public bool RequireHsmBacking
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Explicit key ID to use. If null, one will be generated.
-    /// </summary>
-    public string? KeyId
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Additional metadata to associate with the key.
-    /// </summary>
-    public IDictionary<string, object>? Metadata
-    {
-        get; set;
-    }
-}
-
-/// <summary>
 /// Manages cryptographic keys for wallet operations.
 /// </summary>
 public interface IKeyManager
