+++ src/SdJwt.Net.Wallet/Formats/ICredentialFormatPlugin.cs
@@ -71,100 +71,3 @@ public interface ICredentialFormatPlugin
     /// <returns>True if this plugin can handle the credential.</returns>
     bool CanHandle(string credential);
 }
-
-/// <summary>
-/// Options for parsing credentials.
-/// </summary>
-public class ParseOptions
-{
-    /// <summary>
-    /// Whether to strictly validate the credential structure.
-    /// </summary>
-    public bool StrictValidation { get; set; } = true;
-
-    /// <summary>
-    /// Whether to extract all disclosures.
-    /// </summary>
-    public bool ExtractDisclosures { get; set; } = true;
-}
-
-/// <summary>
-/// Context for creating presentations.
-/// </summary>
-public class PresentationContext
-{
-    /// <summary>
-    /// The audience (verifier) for the presentation.
-    /// </summary>
-    public string Audience { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Nonce from the verifier.
-    /// </summary>
-    public string Nonce { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Key ID to use for key binding JWT.
-    /// </summary>
-    public string KeyId { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Signing algorithm for key binding JWT.
-    /// </summary>
-    public string SigningAlgorithm { get; set; } = "ES256";
-
-    /// <summary>
-    /// Optional state value.
-    /// </summary>
-    public string? State
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Issued at time. Defaults to now.
-    /// </summary>
-    public DateTimeOffset IssuedAt { get; set; } = DateTimeOffset.UtcNow;
-}
-
-/// <summary>
-/// Context for validation.
-/// </summary>
-public class ValidationContext
-{
-    /// <summary>
-    /// Whether to validate the signature.
-    /// </summary>
-    public bool ValidateSignature { get; set; } = true;
-
-    /// <summary>
-    /// Whether to check expiration.
-    /// </summary>
-    public bool ValidateExpiration { get; set; } = true;
-
-    /// <summary>
-    /// Whether to check revocation status.
-    /// </summary>
-    public bool CheckRevocationStatus { get; set; } = false;
-
-    /// <summary>
-    /// Expected issuer.
-    /// </summary>
-    public string? ExpectedIssuer
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Expected audience.
-    /// </summary>
-    public string? ExpectedAudience
-    {
-        get; set;
-    }
-
-    /// <summary>
-    /// Clock skew tolerance.
-    /// </summary>
-    public TimeSpan ClockSkewTolerance { get; set; } = TimeSpan.FromMinutes(5);
-}
